module command registered, doc updates, tests
