+++ tqdm/__init__.py
@@ -3,7 +3,8 @@ from ._tqdm import trange
 from ._tqdm_gui import tqdm_gui
 from ._tqdm_gui import tgrange
 from ._tqdm_pandas import tqdm_pandas
+from ._main import main
 from ._version import __version__  # NOQA
 
 __all__ = ['tqdm', 'tqdm_gui', 'trange', 'tgrange', 'tqdm_pandas',
-           '__version__']
+           'main', '__version__']
